+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.3.0 2021-10-25
+
+- Read `user_id` to provide a permanent ID for the HALO Home account
+
+## 0.2.1 2021-10-23
+
+- Raise HaloHomeError when credentials are not valid
+
 ## 0.2.0 2021-10-21
 
 - Support mesh communication (only need to connect to a single device